requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.2
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
Pillow==8.0.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
virtualenv==20.0.35
